Update new portions on value changed and portions can be fractions

+++ RecEpee/Models/Ingredient.cs
@@ -4,10 +4,10 @@ namespace RecEpee.Models
     public class Ingredient
     {
         public string Name { get; set; }
-        public int? Quantity { get; set; }
+        public double? Quantity { get; set; }
         public string Unit { get; set; }
 
-        public Ingredient GetWithDifferentQuantity(int newQuantity)
+        public Ingredient GetWithDifferentQuantity(double newQuantity)
         {
             return new Ingredient { Name = Name, Unit = Unit, Quantity = newQuantity };
         }
